BeatBuffer Timepoint
now timepoint uses unsigned ints for beat, bar, as well as beats in bar
and beat type.

cleaned up BeatBuffer to use at_bar, at_beat methods of timepoint
added 'const' to the beat buffer getters.

+++ include/timepoint.hpp
@@ -8,36 +8,36 @@ namespace DataJockey {
 			TimePoint();
 			//getters
 			time_type type() const;
-			int bar() const;
-			int beat() const;
+			unsigned int bar() const;
+			unsigned int beat() const;
 			//this simply tells how far into the beat we are.. 0..1
 			double pos_in_beat() const;
-			float beats_per_bar() const;
-			float beat_type() const;
+			unsigned int beats_per_bar() const;
+			unsigned int beat_type() const;
 			double seconds() const;
 			bool valid() const;
 			//setters
 			void type(time_type t);
-			void bar(int val);
-			void beat(int val);
+			void bar(unsigned int val);
+			void beat(unsigned int val);
 			void pos_in_beat(double val);
-			void beats_per_bar(float val);
-			void beat_type(float val);
+			void beats_per_bar(unsigned int val);
+			void beat_type(unsigned int val);
 			void seconds(double val);
 
 			//these reset the internal state
-			void at_bar(int newBar, int newBeat = 0, double newPos = 0.0);
-			void at_beat(int newBeat, double newPos = 0.0);
+			void at_bar(unsigned int newBar, unsigned int newBeat = 0, double newPos = 0.0);
+			void at_beat(unsigned int newBeat, double newPos = 0.0);
 			
 			void invalidate();
 		private:
 			time_type mType;
 			//mirroring jack transport
-			int mBar;
-			int mBeat;
+			unsigned int mBar;
+			unsigned int mBeat;
 			double mPosInBeat;
-			float mBeatsPerBar;	//time signature "numerator"
-			float	mBeatType; //time signature "denominator"
+			unsigned int mBeatsPerBar;	//time signature "numerator"
+			unsigned int mBeatType; //time signature "denominator"
 			double mSeconds;
 	};
 }
